Remove ProcessFile class, fix a few bugs
Due to the way output files work, having ProcessFile as a class
doesn't create a very good interface. Not having the implicit
arguments makes the private interface much nastier, but it is worth
it for the better public interface.
+ This way, processfile can open and close the output file and
  append the package clause instead of the calling function

Several bugs were fixed while verifying the class removal was correct:
+ structs with nested anonymus enums are now generated correctly