Fix typdef extraction
It seems like libclang changed how it represents types in the tree,
with typedefs now falling under the enumerated category. The C++ spec
says they are a special case of enumerated type, so I guess it makes
sense.

Leave the old handling code there in case it changes or someone is
running an older library version.